+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+### Added
+- Add method to get seller by id  in catalog client
 ## [2.6.0] - 2021-02-11
 ### Added
 - Adding RatesAndBenefits Client (https://developers.vtex.com/vtex-developer-docs/reference/benefits)